setup.py works. Changed all import paths to reflect new location of pyc files. Reorganized em/ folder, but it still needs some work.
