+++ em/describe/utilities.py
