+++ em/energy/MMGBSA_CA_L.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python2
 # -*- coding: utf-8 -*-
 """
 Created on Thu Apr 13 23:01:32 2017
