+++ em/in_development/KNN_HMM.py
@@ -1,4 +1,5 @@
-# -*- coding: utf-8 -*-
+#!/usr/bin/python
+
 """
 Created on Fri Jun 10 20:25:03 2016
