+++ em/in_development/align_CLUSTALW2.py
@@ -1,4 +1,5 @@
-# -*- coding: utf-8 -*-
+#!/usr/bin/python
+
 """
 Created on Mon Nov 21 08:53:06 2016
