+++ em/in_development/linker_vector_alg.py
@@ -1,4 +1,5 @@
-# -*- coding: utf-8 -*-
+#!/usr/bin/python
+
 """
 Created on Sun Jan 31 00:16:21 2016
