+++ em/manipulate/Molecular_Rigid_Manipulations.py
